Renamed dirs and files to reference Working Preferences where applicable.
