+++ test/shared/workingPreferencesConverterV2.spec.js
